backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Google Cloud & AI
google-cloud-aiplatform==1.40.0
google-generativeai==0.3.2
google-cloud-secret-manager==2.18.0
google-auth==2.27.0

# Redis & Celery
redis==5.0.1
celery==5.3.6
flower==2.0.1

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1

# WebSocket
websockets==12.0
python-socketio==5.11.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
pendulum==3.0.0

# Integrations
ShopifyAPI==12.5.0
twilio==8.11.1
python-telegram-bot==20.8

# Monitoring & Logging
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
